Minor fixes, optimization, new commands
When shopping for a part you can react with arrow reactions to change part being viewed for up to 5 minutes.

replacecolor command - Replace the background color of the trees using a hex code.

daily command - Adds a number between 50-100 to balance that has a cooldown of one day.

balance command - View balance of yourself or others.